Refactor fixtures into common
