+++ tests/cli/api/dir_create.rs
@@ -4,7 +4,7 @@ use rstest::*;
 use serde_json::json;
 use test_log::test;
 
-use crate::cli::fixtures::*;
+use crate::common::fixtures::*;
 
 #[rstest]
 #[test(tokio::test)]
