+++ tests/cli/api/dir_read.rs
@@ -5,7 +5,7 @@ use rstest::*;
 use serde_json::json;
 use test_log::test;
 
-use crate::cli::fixtures::*;
+use crate::common::fixtures::*;
 
 /// Creates a directory in the form
 ///
