+++ tests/cli/client/fs_metadata.rs
@@ -2,8 +2,8 @@ use assert_fs::prelude::*;
 use predicates::prelude::*;
 use rstest::*;
 
-use crate::cli::fixtures::*;
-use crate::cli::utils::regex_pred;
+use crate::common::fixtures::*;
+use crate::common::utils::regex_pred;
 
 const FILE_CONTENTS: &str = r#"
 some text
