+++ tests/cli/client/fs_read_directory.rs
@@ -4,8 +4,8 @@ use assert_fs::prelude::*;
 use predicates::prelude::*;
 use rstest::*;
 
-use crate::cli::fixtures::*;
-use crate::cli::utils::regex_pred;
+use crate::common::fixtures::*;
+use crate::common::utils::regex_pred;
 
 /// Creates a directory in the form
 ///
