+++ tests/cli/client/system_info.rs
@@ -2,7 +2,7 @@ use std::env;
 
 use rstest::*;
 
-use crate::cli::fixtures::*;
+use crate::common::fixtures::*;
 
 #[rstest]
 #[test_log::test]
