+++ tests/cli/mod.rs
@@ -1,6 +1,4 @@
 mod api;
 mod client;
-mod fixtures;
 mod manager;
 mod scripts;
-mod utils;
