+++ tests/cli_tests.rs
@@ -1,1 +1,2 @@
 mod cli;
+mod common;
