+++ tests/common/utils/predicates.rs
