+++ tests/stress_tests.rs
@@ -2,28 +2,30 @@ use assert_fs::prelude::*;
 use predicates::prelude::*;
 use rstest::*;
 
-mod stress;
+mod common;
+
+use common::fixtures::*;
 
 #[rstest]
 #[test_log::test]
-fn should_handle_large_volume_of_requests() {
+fn should_handle_large_volume_of_requests(ctx: DistantManagerCtx) {
     todo!();
 }
 
 #[rstest]
 #[test_log::test]
-fn should_handle_wide_spread_of_clients() {
+fn should_handle_wide_spread_of_clients(ctx: DistantManagerCtx) {
     todo!();
 }
 
 #[rstest]
 #[test_log::test]
-fn should_handle_abrupt_client_disconnects() {
+fn should_handle_abrupt_client_disconnects(ctx: DistantManagerCtx) {
     todo!();
 }
 
 #[rstest]
 #[test_log::test]
-fn should_handle_badly_killing_client_shell_with_interactive_process() {
+fn should_handle_badly_killing_client_shell_with_interactive_process(ctx: DistantManagerCtx) {
     todo!();
 }
